refactor: split agentic strategy helpers
